Modified service using int and controller.

+++ solar-system/src/main/java/com/galaxy/dao/Dao.java
@@ -8,6 +8,6 @@ public interface Dao<T> {
 	public void add(T t);
 	public void update(T t);
 	public List<T> getList();
-	public T getById(Long id);
+	public T getById(int id);
 	public void remove(int id);
 }
