+++ solar-system/src/main/java/com/galaxy/service/Service.java
@@ -8,7 +8,7 @@ public interface Service<T> {
 	public void add(T p);
 	public void update(T p);
 	public List<T> getList();
-	public T getById(Long id);
+	public T getById(int id);
 	public void remove(int id);
 	
 }
